+++ tests/Unit/Validations/BackwardsCompatible/Equals_Field.php
@@ -1,12 +1,12 @@
 <?php
 
-it('can render equalsfield validation')
-    ->expect(vr()->equalsfield('other_field_name'))
+it('can render equals_field validation')
+    ->expect(vr()->equals_field('other_field_name'))
     ->toEqual('equalsfield,other_field_name')
 ;
 
-it('can render invert equalsfield validation')
-    ->expect(vr()->not->equalsfield('other_field_name'))
+it('can render invert equals_field validation')
+    ->expect(vr()->not->equals_field('other_field_name'))
     ->toEqual('invert_equalsfield,other_field_name')
 ;
 
@@ -15,28 +15,28 @@ $incorrect = ['test' => 'string', 'other' => 'different_string'];
 
 // validate with correct input field
 
-it('can validate equalsfield with correct input', function () use ($correct) {
+it('can validate equals_field with correct input', function () use ($correct) {
     $val = new \Validator\Validator($correct);
 
-    $val->test->equalsfield('other');
+    $val->test->equals_field('other');
 
     expect($val->is_valid())->toBeTrue();
 });
 
-it('can validate equalsfield (not) with correct input', function () use ($correct) {
+it('can validate equals_field (not) with correct input', function () use ($correct) {
     $val = new \Validator\Validator($correct);
 
-    $val->test->not->equalsfield('other');
+    $val->test->not->equals_field('other');
 
     expect($val->is_valid())->toBeFalse();
 });
 
 // validate with incorrect input field
 
-it('can validate equalsfield with incorrect input', function () use ($incorrect) {
+it('can validate equals_field with incorrect input', function () use ($incorrect) {
     $val = new \Validator\Validator($incorrect);
 
-    $val->test->equalsfield('other');
+    $val->test->equals_field('other');
 
     expect($val->is_valid())->toBeFalse();
 });
@@ -44,7 +44,7 @@ it('can validate equalsfield with incorrect input', function () use ($incorrect)
 it('can validate contains (not) with incorrect input', function () use ($incorrect) {
     $val = new \Validator\Validator($incorrect);
 
-    $val->test->not->equalsfield('other');
+    $val->test->not->equals_field('other');
 
     expect($val->is_valid())->toBeTrue();
 });